moved filepreview to another route and component
