+++ src/components/childComponents/CreateAndUpload.js
@@ -1,40 +1,34 @@
 // Create a dataset and upload a file and submit for extraction
 
 import React, {useEffect, useState, useCallback} from 'react';
+import { useNavigate } from "react-router-dom";
 import {useDispatch, useSelector} from "react-redux";
 import LoadingOverlay from "react-loading-overlay-ts";
 import {Box, Button} from "@material-ui/core";
 import Radio from '@mui/material/Radio';
 import RadioGroup from '@mui/material/RadioGroup';
 import FormControlLabel from '@mui/material/FormControlLabel';
-import FormControl from '@mui/material/FormControl';
-import FormLabel from '@mui/material/FormLabel';
 
 import Dropfile from "./Dropfile";
-import Html from "../previewers/Html";
-import Audio from "../previewers/Audio";
-import Video from "../previewers/Video";
-import Thumbnail from "../previewers/Thumbnail";
-import {createEmptyDataset as createEmptyDatasetAction, createUploadExtract} from "../../actions/dataset";
-import {uploadFileToDataset as uploadFileToDatasetAction} from "../../actions/dataset";
-import {checkExtractionStatus, getPreviewResources} from "../../utils/file";
+import {createUploadExtract} from "../../actions/dataset";
+import {checkExtractionStatus} from "../../utils/file";
 import {checkHtmlInDatasetRequest} from "../../utils/dataset";
 import {fetchFilePreviews} from "../../actions/file";
 
 
 export default function CreateAndUpload() {
 	const dispatch = useDispatch();
+	const navigate = useNavigate();
 
 	const [mouseHover, setMouseHover] = useState(false); // mouse hover state for dropzone
-	const [loading, setLoading] = useState(false); // loading overlay state. set to active when dropfile is active
+	const [loading, setLoading] = useState(false); // loading overlay state and button disabled state. set to active when dropfile is active
 	const [loading_text , setLoadingText] = useState("processing"); // loading overlay text.
-	const [previews, setPreviews] = useState([]); // state for file previews
+	const [preview, setPreview] = useState(true); // disabled button state for file preview button
 
 	const listFilePreviews = (fileId) => dispatch(fetchFilePreviews(fileId));
 
 	const datasets = useSelector((state) => state.dataset.datasets);
 	const filesInDataset = useSelector(state => state.dataset.files);
-	const filePreviews = useSelector((state) => state.file.previews);
 
 	const extractor_name = "ncsa.rctTransparencyExtractor"
 
@@ -42,8 +36,6 @@ export default function CreateAndUpload() {
 	const onDropFile = (file) => {
 		setLoadingText("Uploading file");
 		dispatch(createUploadExtract(file, extractor_name));
-		// dispatch(createEmptyDatasetAction(file));
-		// const datasets = useSelector((state) => state.dataset.datasets);
 	};
 
 	// useEffect on filesInDataset for preview generation
@@ -62,6 +54,8 @@ export default function CreateAndUpload() {
 				if (typeof htmlFile.id === "string") {
 					// {"id":string, "size":string, "date-created":string, "contentType":text/html, "filename":string}
 					listFilePreviews(htmlFile.id);
+					setLoading(false); // stop display of Overlay
+					setPreview(true)  // Continue button activated
 				} else {
 					console.log("check html file after 5s");
 					setTimeout(loop, 5000);
@@ -82,21 +76,6 @@ export default function CreateAndUpload() {
 	// TODO how to make this dependency better? Now filesInDataset.id throws an error on start
 
 
-	// useEffect on filePreviews to download preview resources
-	useEffect( async ()=> {
-		if (filePreviews !== undefined && filePreviews.length > 0) {
-			const previewsTemp = [];
-			filePreviews[0].map(async (preview) => {
-				// get all preview resources
-				const preview_config = await getPreviewResources(preview);
-				previewsTemp.push(preview_config);
-				setPreviews(previewsTemp); // set previews
-				setLoading(false); // stop display of Overlay
-			});
-		}
-	}, [filePreviews])
-
-
 	// onDrop function to trigger createUploadExtract action dispatch
 	const onDrop = useCallback(acceptedFiles => {
 		// this callback will be called after files get dropped, we will get the acceptedFiles. If you want, you can even access the rejected files too
@@ -105,62 +84,30 @@ export default function CreateAndUpload() {
 	}, [mouseHover]);
 	// TODO have a dependancy here - mouse hover or dropped file action
 
+	const goToPreviewRoute = () => {
+		let path = '/preview';
+		navigate(path);
+	}
+
 	// We pass onDrop function and accept prop to the component. It will be used as initial params for useDropzone hook
 	return (
 		<Box className="createupload">
-			<LoadingOverlay
-				active={loading}
-				spinner
-				text={loading_text}
-			>
-			<div className="mousehoverdrop" onMouseEnter={()=> setMouseHover(true)} >
-				<Dropfile onDrop={onDrop}
-						  accept={ {'image/png': ['.png'], 'image/jpeg': ['.jpg', '.jpeg'], 'text/html': ['.html', '.htm'], 'text/plain':['.txt']} }
-				/>
-			</div>
+			<LoadingOverlay active={loading} text={loading_text} spinner>
+				<div className="mousehoverdrop" onMouseEnter={()=> setMouseHover(true)} >
+					<Dropfile onDrop={onDrop} accept={ {'text/plain':['.txt']} }/>
+				</div>
 			</LoadingOverlay>
 
-			<FormControl>
-				<FormLabel id="demo-radio-buttons-group-label">Guideline</FormLabel>
-				<RadioGroup row aria-labelledby="demo-radio-buttons-group-label" defaultValue="consort" name="radio-buttons-group">
+			<div className="radio-buttons-group-div align-left">
+				<RadioGroup defaultValue="consort" name="radio-buttons-group" row>
 					<FormControlLabel value="consort" control={<Radio />} label="CONSORT" />
 					<FormControlLabel value="spirit" control={<Radio />} label="SPIRIT" />
 				</RadioGroup>
-			</FormControl>
-
-			<div className="previewBox">
-				{
-					previews.map((preview) => {
-						if (preview["previewType"] === "audio") {
-							return (
-								<div key={preview["fileid"]}>
-									<Audio fileId={preview["fileid"]} audioSrc={preview["resource"]}/>;
-								</div>
-							);
-						} else if (preview["previewType"] === "video") {
-							return (
-								<div key={preview["fileid"]}>
-									<Video fileId={preview["fileid"]} videoSrc={preview["resource"]}/>;
-								</div>
-							);
-						} else if (preview["previewType"] === "thumbnail") {
-							return (
-								<div key={preview["fileid"]}>
-									<Thumbnail fileId={preview["fileid"]} fileType={preview["fileType"]}
-											   imgSrc={preview["resource"]}/>;
-								</div>
-							);
-						} else if (preview["previewType"] === "html") {
-							return (
-								<div key={preview["fileid"]}>
-									<Html fileId={preview["fileid"]} htmlSrc={preview["resource"]}/>;
-								</div>
-							);
-						}
-
-					})
-				}
 			</div>
+			<div className="preview-button align-right">
+				<Button variant="contained" disabled={preview} onClick={goToPreviewRoute}> Continue </Button>
+			</div>
+
 		</Box>
 
 	);
